Add option to read trace file in chunks
Adds trace parser methods to support reading a trace file in in chunks of a specified size rather than one line at a time. The original option still exists as a combination of the read_line and _parse_line methods, but is not currently in use in the main function.